audio: DolbyMS12V2+HEAAC use ms12 HEAAC decoder [1/1]
PD#SWPL-91567

Problem:
use ms12 heaac decoder

Solution:
use ms12 HEAAC decoder
use the skip_amadec mode to develop the PES function of T5D
develop a parser to framing the HEAAC stream

Verify:
t5d

+++ utils/include/aml_android_utils.h
@@ -35,7 +35,7 @@ int aml_getprop_int(const char *path);
 /*
  * Sys Fs Interface
  */
-int aml_sysfs_get_int(const char *path);
+unsigned long aml_sysfs_get_int(const char *path);
 int aml_sysfs_get_int16(const char *path, unsigned *value);
 int aml_sysfs_get_str(const char *path, char *buf, int count);
 int aml_sysfs_set_int(const char *path, int value);
